# lib/api-zod/src/generated/types/userAnalysisHistoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserAnalysisHistoryItem {
  id: number;
  username: string;
  totalTweets: number;
  overallSentimentScore: number;
  analyzedAt: string;
}